Version 3.31.39 (based on 7ec1c3616e6d67efa7bef3124596f321d2800b76)
Performance and stability improvements on all platforms.

Cr-Commit-Position: refs/heads/candidates@{#25314}

+++ src/snapshot-source-sink.h
@@ -24,7 +24,7 @@ class SnapshotByteSource FINAL {
 
   bool HasMore() { return position_ < length_; }
 
-  int Get() {
+  byte Get() {
     DCHECK(position_ < length_);
     return data_[position_++];
   }
